Add 'of' filter for verifying arrays of records.
A common usecase with filtering is that you are given an array of items
(where each item is an array), and you want to verify that each item
adheres to a specification.  Having to do a loop in the code distracts
from what is actually happening and makes it easy to do validation and
processing at the same time rather than validating the full parameter
list first before moving on.

This 'of' filter makes it easy to do validation of records correctly.