Add a new Arrays::flatten filter.
This filter squashes multi-dimensional arrays into a single level.  This
is useful when you have values that may be given as an array or may be
given as a single value and can be nested as such.  Rather than having
to handle recursive checks to see if something is an array or not (or
use something like the RecursiveIteratorIterator), flattening the array
into a single level structure helps a ton.